Tune ambient speed controls and simplify HML LED card behavior.
Lower gradient/breath speed UI ranges, invert chase slider direction for intuitive slow/fast control, switch LED strip default pin to GPIO20, and remove the redundant card-level brightness slider.

Made-with: Cursor

+++ headers/addons/neopicoleds.h
@@ -257,7 +257,7 @@ private:
     // Ambient neopico leds
 	float alBrightnessBreathX;
 	bool alReverse;
-	int alCurrentFrame;
+	float alCurrentFrame;
 	int alFrameToRGB;
 	int alFrameSpeed;
     RGB ambientLight;
